Start of the pointy pop-ups.
