+++ vscode/webviews/OnboardingExperiment.story.tsx
@@ -7,7 +7,7 @@ import { VSCodeStoryDecorator } from './storybook/VSCodeStoryDecorator'
 import { VSCodeWrapper } from './utils/VSCodeApi'
 
 const meta: Meta<typeof LoginSimplified> = {
-    title: 'cody/LoginExperiment',
+    title: 'cody/App-less Onboarding',
     component: LoginSimplified,
     decorators: [VSCodeStoryDecorator],
 }
