+++ vscode/webviews/Popups/InstallCodyAppNotice.module.css
@@ -1,0 +1,86 @@
+.notice {
+    --notice-border-radius: 6px;
+    text-align: left;
+    box-sizing: border-box;
+    padding: 10px 10px 13px;
+
+    background-color: var(--vscode-sideBar-background);
+    color: var(--vscode-sideBar-foreground);
+    border: 1px solid var(--vscode-widget-border);
+    box-shadow: 0 0 8px 2px var(--vscode-widget-shadow);
+    border-radius: var(--notice-border-radius);
+
+    position: absolute;
+    bottom: calc(100% + var(--notice-border-radius));
+    left: calc(-1 * var(--notice-border-radius));
+    width: calc(100vh - 2rem);
+}
+
+.pointy-bit {
+    position: absolute;
+    left: calc(50% - 6px);
+    bottom: 100%;
+    rotate: -45deg;
+    width: 12px;
+    height: 12px;
+    background-color: var(--vscode-sideBar-background);
+    border-left: 1px solid var(--vscode-widget-border);
+    border-bottom: 1px solid var(--vscode-widget-border);
+    box-shadow: 0 0 8px 2px var(--vscode-widget-shadow);
+}
+
+.popup-host {
+    position: relative;
+}
+
+.row {
+    display: flex;
+    flex-direction: row;
+    gap: 4px;
+}
+
+.action-button-container :nth-last-child(-n + 1) {
+    background-color: var(--vscode-button-secondaryBackground);
+    color: var(--vscode-button-secondaryForeground);
+    border: var(--vscode-button-border);
+}
+
+.action-button-container :nth-last-child(-n + 1):hover {
+    background-color: var(--vscode-button-secondaryHoverBackground);
+}
+
+.notice-close {
+    margin-top: -5px;
+    margin-right: -5px;
+}
+
+.notice h1,
+.notice p {
+    all: unset;
+}
+
+.action-button-container {
+    margin-top: 0.6rem;
+}
+
+.notice-text {
+    display: flex;
+    flex-direction: column;
+    gap: 4px;
+    flex: auto;
+    font-weight: 500;
+}
+
+.notice-text h1 {
+    font-weight: bold;
+}
+
+.notice-icon {
+    width: 24px;
+    height: 24px;
+}
+
+/* When there's no notices we don't want the container taking up space and capturing clicks */
+.notice:empty {
+    display: none;
+}
