+++ vscode/webviews/storybook/VSCodeStoryDecorator.module.css
@@ -18,19 +18,61 @@
 /**
  * Define any --vscode-* CSS variables needed. These differ per theme and don't need to be
  * an exact match for any particular VS Code theme. They just need to be good enough.
+ *
+ * For a dark theme, you can use these:
+ * https://sourcegraph.com/github.com/microsoft/vscode/-/blob/extensions/theme-defaults/themes/dark_modern.json
+ *
  */
 
 :root {
-    --vscode-font-family: -apple-system, BlinkMacSystemFont, "Ubuntu", "Droid Sans", "Segoe WPC", "Segoe UI", sans-serif;
-    --vscode-editor-foreground: var(--foreground);
+    --vscode-button-secondaryBackground: #3a3d41;
+    --vscode-button-secondaryForeground: #ffffff;
+    --vscode-button-secondaryHoverBackground: #45494e;
     --vscode-editor-background: #151c28;
-    --vscode-inputValidation-errorBorder: #ff0000;
-    --vscode-inputValidation-errorBackground: #ff000033;
     --vscode-editor-font-size: 13px;
+    --vscode-editor-foreground: var(--foreground);
+    --vscode-font-family: -apple-system, BlinkMacSystemFont, 'Ubuntu', 'Droid Sans', 'Segoe WPC', 'Segoe UI', sans-serif;
+    --vscode-inputValidation-errorBackground: #ff000033;
+    --vscode-inputValidation-errorBorder: #ff0000;
     --vscode-link-color: #7777ff;
-    --vscode-tab-inactiveBackground: #00000033;
-    --vscode-tab-inactiveForeground: #ffffff77;
-    --vscode-tab-activeForeground: #ffffff;
     --vscode-sideBarSectionHeader-border: #ffffff33;
     --vscode-sideBarTitle-foreground: #ffffffcc;
+    --vscode-sideBar-background: #181818;
+    --vscode-sideBar-foreground: #cccccc;
+    --vscode-tab-activeForeground: #ffffff;
+    --vscode-tab-inactiveBackground: #00000033;
+    --vscode-tab-inactiveForeground: #ffffff77;
+    --vscode-widget-shadow: rgba(0, 0, 0, 0.2);
 }
+
+@font-face {
+	font-family: 'codicon';
+	font-display: block;
+	src: url('./codicon.ttf') format('truetype');
+}
+
+:global .codicon[class*='codicon-'] {
+	font: normal normal normal 16px/1 codicon;
+	display: inline-block;
+	text-decoration: none;
+	text-rendering: auto;
+	text-align: center;
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
+	user-select: none;
+}
+
+:global .codicon-modifier-disabled {
+	opacity: 0.5;
+}
+
+:global .codicon-check::before { content: '\eab2' }
+:global .codicon-close::before { content: '\ea76' }
+:global .codicon-gear::before { content: '\eaf8' }
+:global .codicon-link::before { content: '\eb15' }
+:global .codicon-rocket::before { content: '\eb44' }
+:global .codicon-stop-circle::before { content: '\eba5' }
+:global .codicon-terminal::before { content: '\ea85' }
+:global .codicon-thumbsdown::before { content: '\eb6b' }
+:global .codicon-thumbsup::before { content: '\eb6c' }
+:global .codicon-trash::before { content: '\ea81' }
